Fix race condition in temp dir handling.
Set the temp dir location on first usage instead of during compile
time, this prevents a race condition if the process forks after module
compilation.

Also only try to remove the temp dir if the variable has been set, it
will not be if no CGI has been compiled.

This race condition is tested for in t/race-conditions.t, added in
c0a01bc, which now passes.

Fix #24.

+++ lib/CGI/Compile.pm
@@ -167,11 +167,14 @@ sub _build_package {
     return $package;
 }
 
-# we use a tmpdir chmodded to 0700 so that the tempfiles are secure
-my $tmp_dir = File::Spec->catfile(File::Spec->tmpdir, "cgi_compile_$$");
-
+# define tmp_dir value later on first usage, otherwise all children
+# share the same directory when forked
+my $tmp_dir;
 sub _eval {
     my $code = \$_[1];
+
+    # we use a tmpdir chmodded to 0700 so that the tempfiles are secure
+    $tmp_dir ||= File::Spec->catfile(File::Spec->tmpdir, "cgi_compile_$$");
 
     if (! -d $tmp_dir) {
         mkdir $tmp_dir          or die "Could not mkdir $tmp_dir: $!";
@@ -192,7 +195,7 @@ sub _eval {
 }
 
 END {
-    if (-d $tmp_dir) {
+    if ($tmp_dir and -d $tmp_dir) {
         File::Path::remove_tree($tmp_dir);
     }
 }
